pyproject.toml
[tool.poetry]
name = "smatrix"
version = "0.1.0"
description = ""
authors = ["olliecheng <olliecheng@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
wat-inspector = "^0.2.1"
rich = "^13.7.1"
schema = "^0.7.7"

[tool.poetry.scripts]
smatrix = "smatrix.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
